% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractRecords}
\alias{extractRecords}
\title{Extract relevant records from dataframe/tibble of APC data}
\usage{
extractRecords(data, ICDCodes)
}
\arguments{
\item{data}{tibble/dataframe of APC data}

\item{ICDCodes}{ICD codes to include}
}
\description{
extracts data and returns as list of dataframes one for each ICD code
}
