% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadFromDta}
\alias{loadFromDta}
\title{Load Data from stata .dta file}
\usage{
loadFromDta(dtaFile)
}
\arguments{
\item{dtaFile}{path to dtaFile (str)}
}
\description{
imports data and stores in a tibble
}
