% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countOutcomes.R
\name{splitByAge}
\alias{splitByAge}
\title{auxillary function for getAdmissionsTimeSeries()}
\usage{
splitByAge(data, ageBrackets, ageCol)
}
\description{
auxillary function for getAdmissionsTimeSeries()
}
