% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{splitByLocation}
\alias{splitByLocation}
\title{auxillary function for getAdmissionsTimeSeries()}
\usage{
splitByLocation(data, locCol)
}
\description{
auxillary function for getAdmissionsTimeSeries()
}
